Add files via upload
Figure 1 is a bar chart that depicts the number squirrels seen in a sighting. It is interesting to note that the higher the number of squirrels in sightings are, the less prominent they are. This could mean that there are some validity issues with the recorded sightings. Figure 2 is a bar chart that shows the condition of the park at the time of sightings. We can see that the park is more busy than calm, which is not all that surprising due to the location being Central Park. Figure 3 is a bar chart showing the number of sighters during a sighting. From the graph we can see that more often than not, there is only 1 sighter during these sightings. It is also very interesting that the number of sighters does not go over 3 people. Finally, Figure 4 shows the time of day (AM or PM) that the sightings occur. We can see that there are more sightings in the morning than the evening, but it is only a slight difference. This is interesting because it makes you wonder how the time of days could be neck and neck if it is usually darker in the evening. For example, someone could mistaken a squirrel for a raccoon or chipmunks more so in the evening than in the morning.