Add files via upload
I chose box plots for both of my graphs so that understanding and viewing the outliers would be more efficient. Figure 1 is a box plot showing the relationship between the number of squirrel sightings and what time of day the sightings occurred. What I found interesting about this though, was that fact that there were more squirrels seen in the afternoons/evenings than in the mornings. There are also more outliers for the count of squirrels in the mornings as well. Figure 2 is a box plot as well that shows the relationship between how many squirrels were seen, and how crowded the park was at the time. What was interesting about this plot was that there were more outliers in during the busy and calm hours as opposed to the moderate hours.