Milstone 7
Here, I conducted a logistic regression with the number of squirrels sighted as the dependent variable, and other sightings as the independent variable. I did a separate model for both Dogs and Humans to see if one would produce good results. I then did a model with multiple variables from the set to get a better estimate as well. The results were actually very underwhelming. I thought that there would be more correlation between the variables put together being that some columns are based off of one another or at least following the same criteria. I don't think any of the parameter's are good for predictions, so i would recommend none.