Milestone 9
For this milestone, I decided to check the mean number of squirrels sighted and the mean amount of times the squirrel was seen. Being that the other two results were very underwhelming, I decided to go back to the basics and see the of the dependent variable as well as another numerical column. The mean number of squirrels sighted was 4, which is the most normal numerical count I have seen throughout these project results. The next mean is the number of times you see the same squirrel in a sighting, which is 25. I find this mean a bit unreasonable because you can never know for sure if you saw the same squirrel.