Add files via upload
I ran a linear regression model for this milestone with the same variables as I did in milestone 7. I had hoped that since it was a different type of regression, that it would provide different better results. This was not the case and the values are still as insignificant as they were in the last milestone.